In response to Matt's comments.  TBC for CopulaCorrelationMatrixGenerator.random
